tuning and refactoring
